requirements.txt
absl-py==1.0.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
autopep8==1.6.0
black==21.4b2
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.0.0
cycler==0.11.0
Cython==0.29.28
detectron2==0.6
flake8==4.0.1
fonttools==4.31.1
future==0.18.2
fvcore==0.1.5.post20220305
google-auth==2.6.2
google-auth-oauthlib==0.4.6
grpcio==1.44.0
hydra-core==1.1.1
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.4.0
iopath==0.1.9
kiwisolver==1.4.0
Markdown==3.3.6
matplotlib==3.5.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.22.3
oauthlib==3.2.0
omegaconf==2.1.1
opencv-python==4.5.5.64
packaging==21.3
panopticapi @ git+https://github.com/cocodataset/panopticapi.git@7bb4655548f98f3fedc07bf37e9040a992b054b0
pathspec==0.9.0
Pillow==9.0.1
portalocker==2.4.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycodestyle==2.8.0
pydot==1.4.2
pyflakes==2.4.0
pyparsing==3.0.7
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
six==1.16.0
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
toml==0.10.2
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.63.0
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
yacs==0.1.8
zipp==3.7.0
